+++ include/iutest_printers.hpp
@@ -371,7 +371,7 @@ inline void PrintTo(const floating_point<T>& f, iu_ostream* os)
     PrintToFloatingPoint(f, os);
 }
 template<typename T>
-inline void PrintTo(const FloatingPoint<T>& f, iu_ostream* os)
+inline void PrintTo(const internal::FloatingPoint<T>& f, iu_ostream* os)
 {
     PrintToFloatingPoint(f, os);
 }
@@ -413,7 +413,7 @@ inline void PrintToFloat128(const T v, iu_ostream* os)
 }
 
 // NOTE: need libquadmath
-inline void PrintTo(const detail::Float128::Float v, iu_ostream* os)
+inline void PrintTo(const internal::Float128::Float v, iu_ostream* os)
 {
     PrintToFloat128(v, os);
 }