+++ test/printers_tests.cpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2012-2021, Takazumi Shirayanagi\n
+ * Copyright (C) 2012-2022, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -17,6 +17,14 @@
 // include
 #include "iutest.hpp"
 #include "logger_tests.hpp"
+
+struct AlwaysThrow;
+struct Bar;
+struct BigVar;
+struct Hoge;
+struct Point0;
+struct Point1;
+struct Point2;
 
 #if !defined(IUTEST_USE_GTEST)
 
@@ -32,8 +40,9 @@
     (void)(expect); \
     (void)(val)
 
-// unused
-// #define IUTEST_PRINTTOSTRING_CONTAIN(expect, val)
+#define IUTEST_PRINTTOSTRING_CONTAIN(expect, val)   \
+    (void)(expect); \
+    (void)(val)
 
 #endif
 
@@ -93,7 +102,7 @@ IUTEST(PrintToTest, Bar)
 IUTEST(PrintToTest, FloatingPoint)
 {
     ::iutest::floating_point<float> f = 1.0f;
-    ::iutest::detail::FloatingPoint<float> F(1.0f);
+    ::iutest::internal::FloatingPoint<float> F(1.0f);
     IUTEST_ASSERT_STREQ(::iutest::PrintToString(f), ::iutest::PrintToString(F));
 }
 
@@ -124,7 +133,7 @@ IUTEST(PrintToTest, UInt128)
 #if IUTEST_HAS_LONG_DOUBLE
 IUTEST(PrintToTest, LongDouble)
 {
-    ::iutest::detail::LongDouble ld(static_cast< ::iutest::detail::LongDouble::Float>(1.0f));
+    ::iutest::internal::LongDouble ld(static_cast< ::iutest::internal::LongDouble::Float>(1.0f));
     LogChecker ck("1");
     IUTEST_PRINTTOSTRING_CONTAIN(ck, ld);
     IUTEST_STREAMOUT_CHECK(ld);
@@ -134,7 +143,7 @@ IUTEST(PrintToTest, LongDouble)
 #if IUTEST_HAS_FLOAT128
 IUTEST(PrintToTest, Float128)
 {
-    ::iutest::detail::Float128 f128(static_cast< ::iutest::detail::Float128::Float>(1.0f));
+    ::iutest::internal::Float128 f128(static_cast< ::iutest::internal::Float128::Float>(1.0f));
     LogChecker ck("1");
     IUTEST_PRINTTOSTRING_CONTAIN(ck, f128);
     IUTEST_STREAMOUT_CHECK(f128);