+++ test/repeat_tests.cpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2012-2021, Takazumi Shirayanagi\n
+ * Copyright (C) 2012-2022, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -142,18 +142,22 @@ int main(int argc, char* argv[])
 #endif
 {
     MyEnvironment* const env = new MyEnvironment();
-    IUTEST_ASSERT_EXIT( ::iutest::AddGlobalTestEnvironment(env) == env );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::AddGlobalTestEnvironment(env) == env );
     IUTEST_INIT(&argc, argv);
 
-    IUTEST_ASSERT_EXIT( RepeatTestUnspecified() );
-    IUTEST_ASSERT_EXIT( RepeatTestNonFilter(0) );
-    IUTEST_ASSERT_EXIT( RepeatTestNonFilter(2) );
-    IUTEST_ASSERT_EXIT( RepeatTestNonFilter(10) );
+#if defined(IUTEST_USE_GTEST) && GTEST_LATEST
+    ::iutest::IUTEST_FLAG(recreate_environments_when_repeating) = true;
+#endif
 
-    IUTEST_ASSERT_EXIT( RepeatTestWithEmpyFilter(2) );
-    IUTEST_ASSERT_EXIT( RepeatTestWithEmpyFilter(3) );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestUnspecified() );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestNonFilter(0) );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestNonFilter(2) );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestNonFilter(10) );
 
-    IUTEST_ASSERT_EXIT( RepeatTestWithFilter(3) );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestWithEmpyFilter(2) );
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestWithEmpyFilter(3) );
+
+    IUTEST_TERMINATE_ON_FAILURE( RepeatTestWithFilter(3) );
 
     printf("*** Successful ***\n");
     return 0;