+++ test/set_up_failure_tests.cpp
@@ -61,10 +61,10 @@ int main(int argc, char* argv[])
 #endif
     const int ret = IUTEST_RUN_ALL_TESTS();
 
-    IUTEST_ASSERT_EXIT( ret != 0 );
-    IUTEST_ASSERT_EXIT( ::iutest::UnitTest::GetInstance()->failed_test_count() == 2 );
-    IUTEST_ASSERT_EXIT( setup_flag == 0 );
-    IUTEST_ASSERT_EXIT( teardown_flag == 1 );
+    IUTEST_TERMINATE_ON_FAILURE( ret != 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::UnitTest::GetInstance()->failed_test_count() == 2 );
+    IUTEST_TERMINATE_ON_FAILURE( setup_flag == 0 );
+    IUTEST_TERMINATE_ON_FAILURE( teardown_flag == 1 );
 
     printf("*** Successful ***\n");
     return 0;