+++ test/shuffle_tests.cpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2012-2021, Takazumi Shirayanagi\n
+ * Copyright (C) 2012-2022, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -70,11 +70,15 @@ int main(int argc, char* argv[])
 #endif
 {
     MyEnvironment* const env = new MyEnvironment();
-    IUTEST_ASSERT_EXIT( ::iutest::AddGlobalTestEnvironment(env) == env );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::AddGlobalTestEnvironment(env) == env );
     IUTEST_INIT(&argc, argv);
 
     ::iutest::IUTEST_FLAG(repeat) = 3;
     ::iutest::IUTEST_FLAG(shuffle) = true;
+
+#if defined(IUTEST_USE_GTEST) && GTEST_LATEST
+    ::iutest::IUTEST_FLAG(recreate_environments_when_repeating) = true;
+#endif
 
     const int ret = IUTEST_RUN_ALL_TESTS();
     if( ret != 0 ) return ret;